Add files via upload
This Power BI dashboard provides an in-depth analysis of the Indian Premier League (IPL) from 2021 to 2024, highlighting key performance indicators (KPIs) for both batting and bowling statistics. It offers a detailed overview of player performances, team comparisons, and overall trends, with specific emphasis on individual and team achievements.
Purpose:
The IPL Performance Dashboard is designed to provide actionable insights into the league's performance dynamics. It aids in strategic decision-making for team management, enhances the viewing experience for fans, and celebrates individual and team accomplishments. By presenting a comprehensive view of batting and bowling statistics, along with detailed player and team analyses, this dashboard serves as an essential tool for anyone involved in or following the IPL.